Починила окончательно заявки

+++ src/main/java/ru/praktika/kotouslugi/controller/KotopravaController.java
@@ -4,10 +4,8 @@ import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.praktika.kotouslugi.model.KotopravaRequisition;
-import ru.praktika.kotouslugi.model.Requisition;
 import ru.praktika.kotouslugi.model.response.BaseResponse;
 import ru.praktika.kotouslugi.service.KotopravaService;
-import ru.praktika.kotouslugi.service.RequisitionService;
 
 import java.util.List;
 import java.util.Map;
@@ -44,7 +42,7 @@ public class KotopravaController extends BaseController{
             @ApiResponse(code = 200, message = "OK", response = BaseResponse.class),
             @ApiResponse(code = 401, message = "Не авторизованный пользователь"),
             @ApiResponse(code = 500, message = "Внутренняя ошибка") })
-    public BaseResponse<Long> createPravaRequisition(@ApiParam(name = "request", value = "Тело запроса", required = true)
+    public BaseResponse<Integer> createPravaRequisition(@ApiParam(name = "request", value = "Тело запроса", required = true)
                                                    @RequestBody Map<String, Object> request) {
         return wrapper((s) -> kotopravaService.createPravaRequisition(request));
     }
