+++ src/main/java/ru/praktika/kotouslugi/dao/KotopravaRepository.java
@@ -3,5 +3,5 @@ package ru.praktika.kotouslugi.dao;
 import org.springframework.data.repository.CrudRepository;
 import ru.praktika.kotouslugi.model.KotopravaRequisition;
 
-public interface KotopravaRepository extends CrudRepository<KotopravaRequisition, Long> {
+public interface KotopravaRepository extends CrudRepository<KotopravaRequisition, Integer> {
 }
